+++ mailpoet/assets/js/src/email-editor/engine/components/styles-sidebar/screens/screen-typography.tsx
@@ -1,0 +1,12 @@
+import { __ } from '@wordpress/i18n';
+import { DimensionsPanel } from '../panels/dimensions-panel';
+import { ScreenHeader } from './screen-header';
+
+export function ScreenTypography(): JSX.Element {
+  return (
+    <>
+      <ScreenHeader title={__('Typography', 'mailpoet')} />
+      <DimensionsPanel />
+    </>
+  );
+}
